backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pyotp==2.9.0
sqlparse==0.5.5
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
